Fix a bug in SimpleSMTLIBSolver.Dispose() exposed by pressing CTRL+C. Z3 and Symbooglix get the signal so Z3 exits but Symbooglix continues and then at some point will try to call kill on an already terminated process.

+++ src/Symbooglix/Solver/SimpleSMTLIBSolver.cs
@@ -145,7 +145,9 @@ namespace Symbooglix
                     Console.WriteLine("Cancelled reading stderr");
                     TheProcess.CancelOutputRead();
                     Console.WriteLine("Cancelled reading stdout");
-                    TheProcess.Kill();
+                    if (!TheProcess.HasExited)
+                        TheProcess.Kill();
+
                     Console.WriteLine("killed process");
                     lock (ComputeSatisfiabilityLock)
                     {
